Create ConfigurationWithSecrets
