+++ src/common/bootstrap/configuration/pipeline/embedding/datasources/datasources_configuration.py
@@ -1,8 +1,11 @@
+from abc import ABC
 from enum import Enum
-from typing import Any, Literal, Optional, Union
+from typing import Literal, Optional, Union
 
-from pydantic import BaseModel, ConfigDict, Field, SecretStr
+from pydantic import ConfigDict, Field, SecretStr
 from pydantic_settings import BaseSettings
+
+from common.bootstrap.secrets_configuration import ConfigurationWithSecrets
 
 # Enums
 
@@ -54,23 +57,13 @@ class PdfSecrets(BaseSettings):
 # Configuration
 
 
-class DatasourceConfiguration(BaseModel):
+class DatasourceConfiguration(ConfigurationWithSecrets, ABC):
     name: DatasourceName = Field(
         ..., description="The name of the data source."
     )
     export_limit: Optional[int] = Field(
         None, description="The export limit for the data source."
     )
-
-    def model_post_init(self, context: Any) -> None:
-        self.secrets = self.get_secrets(secrets_file=context["secrets_file"])
-
-    def get_secrets(self, secrets_file: str) -> BaseSettings:
-        secrets_class = self.model_fields["secrets"].annotation
-        secrets = secrets_class(_env_file=secrets_file)
-        if secrets is None:
-            raise ValueError(f"Secrets for {self.name} not found.")
-        return secrets
 
 
 class ConfluenceDatasourceConfiguration(DatasourceConfiguration):
